+++ phpunit_mysqli.xml
@@ -8,6 +8,11 @@
     </php>
     <testsuite name="mysqliLoadedTest">
       <file>tests/mysqliLoadedTest.php</file>
+    </testsuite>
+    <testsuite name="mysqliMakeDB">
       <file>tests/mysqliMakeDB.php</file>
+    </testsuite>
+    <testsuite name="mysqliConnectionTest">
+      <file>tests/mysqliConnectionTest.php</file>
     </testsuite>
 </phpunit>
